Add whisper action

+++ actions/facedetect/src/main/java/io/metaloom/cortex/action/facedescription/FaceDescriptionModel.java
@@ -24,7 +24,7 @@ public enum FaceDescriptionModel implements LargeLanguageModel {
 	}
 
 	@Override
-	public int contextWindow() {
+	public long contextWindow() {
 		return 4096;
 	}
